# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "nimir",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/annotations/data/{data_id}": {
      "get": {
        "tags": [
          "annotations"
        ],
        "summary": "Read Annotations By Data",
        "description": "获取某个数据的标注",
        "operationId": "read_annotations_by_data",
        "parameters": [
          {
            "name": "data_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Data Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnnotationOut"
                  },
                  "title": "Response Annotations-Read Annotations By Data"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "annotations"
        ],
        "summary": "Update Data Annotations",
        "description": "更新数据的标注",
        "operationId": "update_data_annotations",
        "parameters": [
          {
            "name": "data_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Data Id"
            }
          },
          {
            "name": "processing_stage",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "处理阶段",
              "title": "Processing Stage"
            },
            "description": "处理阶段"
          },
          {
            "name": "workflow_execution_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workflow Execution Id"
            }
          },
          {
            "name": "node_execution_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Node Execution Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AnnotationCreate"
                },
                "title": "Annotations"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnnotationOut"
                  },
                  "title": "Response Annotations-Update Data Annotations"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "annotations"
        ],
        "summary": "Delete Data Annotations",
        "description": "删除数据的所有标注",
        "operationId": "delete_data_annotations",
        "parameters": [
          {
            "name": "data_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Data Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Annotations-Delete Data Annotations"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/annotations/workflow/{workflow_execution_id}/{node_id}": {
      "get": {
        "tags": [
          "annotations"
        ],
        "summary": "Get Annotations By Workflow Node",
        "description": "获取工作流特定节点的数据标注",
        "operationId": "get_annotations_by_workflow_node",
        "parameters": [
          {
            "name": "workflow_execution_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Workflow Execution Id"
            }
          },
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Node Id"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  },
                  "title": "Response Annotations-Get Annotations By Workflow Node"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/annotations/project/{project_id}/stage/{stage}": {
      "get": {
        "tags": [
          "annotations"
        ],
        "summary": "Get Annotations By Stage",
        "description": "获取项目特定处理阶段的标注",
        "operationId": "get_annotations_by_stage",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          },
          {
            "name": "stage",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Stage"
            }
          },
          {
            "name": "workflow_execution_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workflow Execution Id"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  },
                  "title": "Response Annotations-Get Annotations By Stage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnnotationCreate": {
        "properties": {
          "frontend_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frontend Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "points": {
            "type": "string",
            "title": "Points",
            "default": ""
          },
          "color": {
            "type": "string",
            "title": "Color",
            "default": "#FF0000"
          },
          "label_id": {
            "type": "integer",
            "title": "Label Id"
          },
          "data_id": {
            "type": "integer",
            "title": "Data Id"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "labelme_data": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Labelme Data"
          },
          "processing_stage": {
            "type": "string",
            "title": "Processing Stage",
            "default": "original"
          },
          "workflow_execution_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Execution Id"
          },
          "node_execution_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Node Execution Id"
          },
          "annotation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Annotation Id"
          }
        },
        "type": "object",
        "required": [
          "type",
          "label_id",
          "data_id",
          "project_id"
        ],
        "title": "AnnotationCreate"
      },
      "AnnotationOut": {
        "properties": {
          "frontend_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frontend Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "points": {
            "type": "string",
            "title": "Points",
            "default": ""
          },
          "color": {
            "type": "string",
            "title": "Color",
            "default": "#FF0000"
          },
          "label_id": {
            "type": "integer",
            "title": "Label Id"
          },
          "data_id": {
            "type": "integer",
            "title": "Data Id"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "labelme_data": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Labelme Data"
          },
          "processing_stage": {
            "type": "string",
            "title": "Processing Stage",
            "default": "original"
          },
          "workflow_execution_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Execution Id"
          },
          "node_execution_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Node Execution Id"
          },
          "annotation_id": {
            "type": "integer",
            "title": "Annotation Id"
          },
          "created": {
            "type": "string",
            "format": "date-time",
            "title": "Created"
          },
          "modified": {
            "type": "string",
            "format": "date-time",
            "title": "Modified"
          }
        },
        "type": "object",
        "required": [
          "type",
          "label_id",
          "data_id",
          "project_id",
          "annotation_id",
          "created",
          "modified"
        ],
        "title": "AnnotationOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        }
      }
    }
  }
}